requirements.txt
agno
ollama
openai
sqlalchemy
pyngrok
python-dotenv
requests
pyTelegramBotAPI
